(BWDIC!) Add `unalternates', drop *alternates* + "alternates -" idiotism..
The "alternates -" was a quick+dirty hack to be able to unset the
alternates list, but a regular `unalternates' command is surely
the much better alternative.

Also change the behaviour of `alternates' and let it by default
append to the list of alternates, only replace the current list if
we are running in *posix* mode.

While here, drop the string list that we used to use for
alternates, instead embed alternates in our normal hashmap lookup
mechanism of nam-a-grp.c.  It is a little more expensive at
creation and unset time, but the lookup may even improve it a bit.
With the future dynamic hashmaps it will be the superior approach.

Luckily i did this, i detected a bug present since nail 10.00 as
of September 2002, preventing *allnet* from functioning.  Add
a test for this.

And drop the *alternates* variable, it is now much too expensive
to maintain with the hashtable approach, and also unnecessary:
simply support the `vput' modifier for `alternates' instead!