FIX [81d7f4d8] (IMAP: try (hard) to reinstantiate..) (Paul Vojta)..
Another fallout of IMAP reinstantiation.
We have extended support for proto:// prefixes so that users can
now say, e.g., "? copy 1 maildir:///tmp/nonexistent.mdir" and that
thing will be newly created as necessary, and also the operation
will fail shall it exist but not be something that looks like
a valid Maildir.

Of course all this is a hack, because i refrained from doing it
right from the start and add global bypasses and carriers, but
continued to use the brain-damaged "pass strings all through, tear
them apart and join them as necessary" even though i knew from the
start that this looses information.  We should: parse user input,
create objects, and keep their instances as long as the operation
they are involved in lingers around.  Here: URL.  Anyway.

This error now was that we splitted off the protocol (Paul's
example was imaps://) first, followed by simply reassigning
imap:// because we did not know any better afterwards, and that in
turn results in a "is-same-host" mismatch because of the protocol
prefix (which is also wrong, but especially so if we would degrade
an already secure channel, but that is a different story).