FIX: OH!  The codebase NEVER supported *mime-encoding*=base64!
This is the total crap blues.
I should have listened to my stomach and introduce the global
bypasses, turning them to function argument carriers over the
years.  And nothing else but model-view-controller and of course
filter chains and event-driven software is the king.
We not yet all of that, in practice.  Sigh.