n_idec_buf(): adjust errors for optional binary/hexadecimal prefixes..
After some error in the tor software (but which likely could not
really do better with the standard functions) all libraries (which
had to) turned to an interpretation of the C standard which says
that the prefix may optionally precede an otherwise valid
sequence, which means that "0x" is not a STATE_INVAL error but
gives a "0" result with a "STATE_BASE" error and a rest of "x".

I never did it that way, but falsely remembered that the C++
library does so if the base is 0 not explicit 16.  No.
Follow that, i finally think it is the better behaviour.