n_mimetype_classify_file(): tweak/fix message/rfc822 C-T-E detection..
The base of the chosen Content-Transfer-Encoding: depends on
*mime-encoding*, but for message/rfc822 RFC 2046 explicitly
restricts the allowed c-t-e to 7-bit, 8-bit and binary.  Because
of a test too simple we would choose binary for quoted-printable
and base64 instead of falling back to the non-black-box ones.

Of course we need a flat parser which is capable to re-encode all
the possible nodes (parts) of the tree (message).