Temporarily freeze variables set via -S..
This was long in TODO, so it must be said that i have doubts since
any update or change is silently rejected.  This is so to not
interfere with *errexit* or *posix*: in $POSIXLY_CORRECT mode any
normal startup with -S becomes a problem otherwise.

On the other hand it really is a huge problem unless we have
a functional `local' command modifier so that 'local set i' and
-Si=100 cannot clash!

But we need it now to avoid initialization order issues for
*ssl-config-file*: the command line should succeed anyhow!