{make-{config,emerge},cc-test}.sh: no command -v (Norman, Ralph, Stéphane)..
Norman Ramsey reported

  I'm building s-nail from source.  But in the configuration step,
  it mistakenly identifies /home/nr/bin/clang as a C compiler.
  This script does not have its executable bit set, and therefor
  it should not be treated as a candidate C compiler.

and Ralph Corderoy responded to my reply

  $ >/tmp/foo
  $ PATH=/tmp:$PATH bash -c 'command -v foo'
  /tmp/foo

and it turns out the POSIX wording my be ambiguous, at least bash
(unless POSIXLY_CORRECT) and dash (always) mistreat command -v in
that they return a non-executable thing.  To workaround this,
implement something on our own.

In a corresponding POSIX issue Stéphane Chazelas pointed out that
there may be empty path components and that the sh(1) noglob
parameter needs to be turned off in order to avoid according
expansion in "set -- $PATH"!

Weeks later i finally realized that using

  : ${varname:=`shellfun bla varname`}

is not portable and changed all of it again.