FIX iconv for main body part (since EVER!) (Doug McIlroy, Random832)..
On the TUHS list Doug McIlroy responded to a mail of mine and
commented on faulty character set conversion artifacts, and it
turns out that the codebase has never been truly prepared for
faulty character set conversions regarding destination character
sets (non-convertible characters).  Never.

Getting this going before the v15 filter-based rewrite is pretty
much hacky in sofar as the topmost level should create a filter
chain for the exact purpose (send-to-display, -file, -mbox, or
create new message) which can the react truly accordingly.  This
could also involve the user in interactive mode by asking him
whether something should be reencoded in the -to-file cases, for
example.  Not yet, that.