FIX sigsuspend(2) race condition (since EVER! Only seen on OpenBSD)..
It seems on OpenBSD singleprocessor the child starts running
first, whereas on other OSs the parent continues execution for
a while.  The codebase did not account for the former, we need to
explicitly block SIGCHLD before the fork() and (thus) unblock it
in the parent and (thus also) the child to avoid that a dying
child causes SIGCHLD before our parent has assigned the PID to the
management structure.