FIX debug assertion (Thomas Dickey)..
It turns out that [bf908249] (a_crese_polite_rt_mft_move(): if
upgrading all to To: - simply redo!) was _too simple_ instead,
because it forgot to update the type of the node if we do not
redoing the list, but simply assigning to To:, triggering a debug
assertion later on caused by Cc: entries ending up in To: but
nonetheless proclaiming they were in Cc:.

One more such missing-concentration errata, but also one more
because of the terrible way we do things.  These nodes only need
to know where they reside because they become joined into a single
list for easier iteration and then splitted into their sublists
over and over again.  That is not the way this should be, even
though that is the historic approach.