FIX i_strdup(), broken since [354fc47e] as of 2017-07-16..
[354fc47ec4b6d209a66432894e348f9db8235be4] (strings.c: tweak
various old char* functions, 2017-07-16) actually causes
the last character of a non-empty mail address string (plus) to be
overwritten by NUL.

This does not break URLified S/MIME *smime-sign-cert-* etc.
lookups because that truncation happens for `set' as well as for
any lookup.  It could cause ambiguities for addresses etc. which
are identical except for the very last character.