Improve *mime-counter-evidence* deep inspection (Doug McIlroy)..
The deep application/octet-stream content inspection that is
controllable by bit 4 (0b1111) did yet use the normal MIME file
classification mechanism, which does not allow any content with
control characters other than \t, \n and such to be treated as
text.

It happened that Doug McIlroy has sent a mail with a single
embedded DC4 (device control 4) in a 2099 bytes message body.
In order to be able to easily look into such a message in the
future this changeset clears the way to pass through the actual
sendaction to the MIME classifier, and henceforth
SEND_TODISP(_ALL)? as well as SEND_QUOTE(_ALL)?, which mangle any
text to mask control characters etc., will wave through parts with
a certain (TODO fixed) number of control characters.