FIX [0ede309c] as of 2016-12-26! (Ralph Corderoy)..
[0ede309c] (Rudely fix the RFC 5322 dot-atom, the "Dr. Problem"
(Dr. Werner Fink), 2016-12-26) apparantly broke generation of all
address-related fields with high bit bytes because it started
calling mime_write_tohdr_a() for the non-address related parts of
such a header instead of mime_write_tohdr()!

Thanks a lot to Ralph Corderoy for reporting this issue!!

Grazy, but we actually have no tests which test this!
I have introduced quite a lot of tests accompanying the set of
changes back then, e.g., for `addrcodec', but it all boils down to
my usual and in the meantime well-known problem of this codebase
which i complain about since the very first day i work on it, and
that is that it does not know about objects but combines and tears
apart and combines and tears apart and forgets a lot of context
all along the way!

Here it should be an "Address".  We do test
Address::dump_to_display() quite a lot, since `addrcodec' allows
direct access to this functionality, but we do not test
Address::dump_to_wire() because that is not the way it works.  We
need to generate a complete message in order to test that.  Until
the healing that comes with v15.

For S-roff i will do things differently and FIRST convert it to
what i think is correct and THEN do anything else!  Is that also
wrong..

Since i am repeating myself oh so often, do it constructively at
least once:

Thanks a lot to Ralph Corderoy for reporting this issue!!