Patch configure script to not override build variables..
Remarks: Steffen has adjusted the patch to use -n instead of
'"x$x" = x', as -n is used anywhere in the codebase.
(-n was in V8 UNIX usr/bin/test.c and at least starting with
4.1BSD in bin/test.  The [(1) operator came a bit later, but
1992-06-03 (CSRG, committer Elan Amir) is probably soon enough to
be acceptable for S-*.)

Paride needs this because even an empty LD_LIBRARY_PATH will be
perpetuated in the configuration and can then (at least portably)
no longer be overwritten from the command line due to the
recursive make approach we (need to) use.