Do not bake kernel version into binary (Simon McVittie, Paride Legovini)..
I wrote to debian-devel@ due to the reproducibility problems
(always ok for x86_64, mess with the others), and Simon responded
with the absolutely right fact that the kernel version really has
no place in the baked in identification.  Go for uname(1) -sm
instead of -srm.

Also clarify that $OSFULLSPEC is not baked into the binary.