n_folder_mbox_prepare_append(): FIX invalid MBOXes (broken by [90861930])..
RFC 4155 specifies that each record in an MBOX file must be
terminated with an completely empty line.  Apparantly there is
still software out there which does not place this trailing
newline when appending messages to MBOX files, and since i have
added n_folder_mbox_prepare_append() in commit [90861930] last
year we do no longer handle this case due to a missing switch()
fallthrough.
Instead do it all differently and require a read of 1 for files of
1 byte, and 2 for files of 2 bytes or more.  And thus handle those
boxes correctly again.