cc-test.sh: add test for [f4db93b3]; more MBOX mishandling, everwhere..
Sigh.  Well, one more reason to use model-view-controller and
objects in general.  When doing "copy SOME-MBOX" and the message
to be copied did not have a trailing newline, we would mess the
target MBOX.  Etc.

So audit the codebase for all possibilities which possibly result
in storing data in MBOX files (sic), and fix the spots found.
Like i said, this needs to be a VFS superobject with a MBOX
subclass that has .open(), .close(), .append(), .insert() or
whatever, then each mailbox type can do whatever is appropriate.
But that is future work, still.

While here, drop the historic "config.h:#define APPEND".
And add some tests for invalid MBOX handling as well as copying
over also those with *mbox-rfc4155* set.