make.rc,make-config.sh: use comma- not space-separated mulchoice..
BSD make fails to pass through a space-separated value to recursive
make invocations and shells, so introduce comma-separation and
use it exclusively, continue support for space separation, though.

While here case normalize mulchoice user values: case-insensitive
matching was used for the validity check, but not for the later
$IFS based value splitting that finally was used to implement
the splitting.  v14.9.10 bug...