Add *forward-inject-tail*, *quote-inject-{head,tail}*, `~Q'..
and extend the meaning of *quote*, too: if that yet would produce
the string "FROMHEADER wrote:\n\n" we will look for the newly
introduced latter two variables.  And as below.

All of *{forward,quote}-inject-{head,tail}* will be passed through
a newly introduced formatter function, a_coll_quote_message().
This will introduce a compose-mode specific set of formats,
getting more and better as time goes by.
The generated output honours *quote-fold*.

This changes the meaning of *forward-inject-head* which has been
introduced in v14.9.0 in that it will also be passed through the
formatter.

While here, introduce the new command escape `~Q' that performs
full *quote* cycles on the given message list.