a_message_match_sender(): fix "matches first only"; drop UUCP stuff..
It turns out that all this mess never really worked.  The problem
is that name1() and nameof() return an entire
From:/Sender:/whatever line, but at least the former can of course
consist of multiple distinct network addresses.  Yet, the return
value was used as if it would be only one, and skin()ned, for
example.  At other places only one would have been used, whatever.
This is not as bad as it sounds in effect, because of our terribly
bad address parsers and the first TODO note i have ever written
(in 2012), we simply use the strings anyway, and tear them apart
and join them again so and so often.

You could search fulltext "? ', ADDRESS'", for example, and it
would match.  Now fixed.

So drop the minimal UUCP address support somewhat, and merge
name1() and nameof() in n_header_senderfield_of(), document that
the return value needs to become lextract()ed.

Actually _do_ lextract() the return value at some places, for
example a_message_match_sender().