Support Fcc: headers in -t'emplates and editor..
the nice thing of which is that the entire body of the header
forms a single target, whereas otherwise mysterious quoting rules
apply when doing stuff in To:/Cc:/Bcc:.
(Unfortunately we do not use shell quoteing rules yet.  The
question is whether that will be possible at any time in the
future, due to backward-compatibility issues.)

Fcc: is target to *exandaddr* just as the mentioned.

While here, allow more targets, like imap:// should work i guess
(untested).