Do not process From: alternates if Sender: is set (Michael Dressel)..
Another point that i did not get right at first, Michael pointed
out all he wants is really going for what he stated, and put some
receivers in Cc:.  It turns out Heirloom and we always took into
account From: in the alternates list automatically, even if Sender:
is explicitly set.

This may even not be false as such, the real problem i have introduced,
and that is that sending performs alternates processing just as does
replying, but only the latter is defined by the standard.
I did not see a real difference in respect to alternates, however.
But according changes cannot be cherry-picked to master, anyhow