Fix (BWDIC!) `~^ header show': optionally show address type in field 1..
The problem is that pipe specifications and file names cannot be
bounded otherwise, we put the raw value, it is not encoded nor
shell-quoted.  Sic.

So we need a backward-incompatible change, and output the actual
address type in field one if showing files, pipes or names
(aliases).  Network email addresses do not change format.