MLE: n_tty_addhist(): split out core to a_tty_hist_add()..
Doing so allows us to avoid setting n_PS_ROOT bit during loading
of the history file, because we now do no longer evaluate all the
bells and whistles that detect whether an entry is to be added for
real, or not.  This also avoids looking up *line-editor-disable*
and (optionally) *history-gabby* for each history entry, as well
as all the system calls made to block signals.  Uff!!