(BWDIC!) VAL_PRIVSEP_USER -> VAL_DOTLOCK_PS_USER (and helper, too)..
The original idea was to have a very small privilege separated
child which could be used for multiple small things which require
higher privileges, for example, dotlock files, $MAIL MBOX creation
with appropriate privileges, etc.

This assumption was false, if we will ever reach the point in time
where more functionality is needed, separate specific helpers will
be created as necessary.