Auto-fix when MBOX had From_ errors on read (Dr. Werner Fink)..
When we sendmp() the message out we now also look whether we had
seen MBOX From_ errors while reading in the MBOX.
If so, do not require *mbox-rfc4155* to perform MBOXO quoting, do
it automatically.

(TODO In v15 the Message::Part::to_wire() will recognize its
contents need to be reencoded anyway.  Then this will be
graceful.)