Disable usage of O_NOFOLLOW again (Russell Bell)..
Mr. Russell Bell reported that he could not `save' to a file.  It
turned out that an intermediate target was a symbolic link, and we
explicitly disallowed doing so in [10990e4c] in August 2016.

But since in our case these opens affect only files which belong
to the user a.k.a. since we do not change ownership nor access
mode of those files after we eventually have created them anew,
simply drop usage of O_NOFOLLOW again.
Mr. Bell has found a bug.