`mimeview': fix binary formats (Russell Bell)..
Well, i _had_ tested that, but it takes a long time until
something goes out, and rebasing and such, it can happen.
Here it seems we made text-format handlers finally work, and
forgot the capability to use binary ones along the way.

  I. We need the v15 filter based approach, that is the _only_ sane
     way to act upon mail data.
 II. Test will follow.
III. No, i usually do not start MIME viewers which are not
     copiousoutput myself.  If something like that really enters,
     it is most often interesting enough to save it somewhere
     first.