FIX IMAP searches via IMAP without matches (Dirk-Wilhelm Peters)..
Dirk-Wilhelm reported that using IMAP search expressions via real
IMAP always returns a matching message even if it should not!

It turned out this was a bug introduced in [08799869] (First
(rather small) step in reworking getmsglist().., 2016-10-02),
a not-so-small commit made _over_two_years_ago_!!!

I simply missed to also convert the native IMAP side to the new
paradigm, which is returning the number of matching messages
instead of only a success status -- and since even a non-matching
but otherwise correct IMAP search is a "success", we had this
problem.

What a terrible bug.
Thank you, Dirk-Wilhelm!