RAND_DRBG_set_reseed_defaults() (Dr. M. St. Pierre, Kurt Roeckx)..
It turns out that OpenSSL 1.1.1 and above changed the PRNG
implementation.  The PRNG is now splitted in master pool and two
servers, one of which feeds the RAND_bytes() we use, and is fed by
the master pool.  This master reseeds itself via /dev/urandom /
getrandom / whatever native there is, and this can fail the same
way as the initial seeding can.

Therefore our (somewhat) carefully crafted _xtls_rand_init(),
which looks out for RAND_status() unless OpenSSL reports
a successul initial seeding, is obsolete, or even invalid, as each
and every RAND_bytes() can now (in theory) cause a reseed, and
fail!

Since we do not have such strict requirements on the actual
quality of the PRNG (though the default should be plenty, should
it), Dr. St. Pierre suggested changing the OpenSSL strategy.
Therefore we disable reseeding if we find the prototype for the
corresponding function.

An alternative approach would have been to keep our own random
implementation even in the OpenSSL case, and possibly use OpenSSL
only for seeding of our own generator.  Maybe we do this later,
when we have converted it to a ChaCha20 or so algorithm, covered
by a Blake2 digest.