a_dotlock_create(): tweak dotlock-ps (Jean-Marc Pigeon, Russell Bell)..
Jean-Marc reported back in November that gcc reports that the
snprintf(3) may become truncated (a week later Russell Bell
reported the same).

Whereas this would do no harm (as the result is de-facto
a somewhat random permutation of whatever input we get, which is
exactly what we want), finally change this after Jean-Mark pointed
it out again.  The actual fix is surprisingly complicated, and
will require a follow-up commit, too.
But with it, it should now work out really fine.

(And i add that the origin of this piece of code is pretty old and
more or less comes from NetBSD Mail.)