(BWDIC!) Tweak addr parser a bit, add GNOT_A_LIST,GNULL_OK (Dr. Werner Fink)..
Any argument given to the command line and such will undergo
a primitive parse step and may be interpreted as a list of
addresses instead; e.g.

  -b 'a@b d@e'

gives two receivers in one go.
We yet had no mechanism to enforce saying "the users means _one_
address", except for the hack that has been introduced to allow
the `addrcodec` command to become implemented.

Also, nalloc() _must_ return a valid object, the only possibility
which exists is to flag it invalid, but NIL is not allowed.

All this is bad.  Add two new flags, GNOT_A_LIST to state that we
really want the argument to become interpreted as a single
addressee, and GNULL_OK to mean the obvious.

Change `addrcodec` as well as -r (this is the incompatibility) to
make use of that.  Change nalloc() to always use the possibly
overworked input/output string, rather than insisting on using the
user argument as given.  Together with "tweaks" in the terrible
addrspec_with_guts() skinner thing this allows for things like

  ? addrc e du , e <w@d> , d
  "du , e , d" <w@d>

instead of failing with INVAL error.