Reintroduce ^From_ MBOXO with *mime-encoding*=8b (too many!)..
[cacee1b5] (Rewrite file-content classification.., 2013-01-02)
removed S-nail's capability of generating MBOXO quoted messages.

It was planned to readd this back then, in that the MIME
classifier would have (will, somewhen) generate status info with
hint results, and let the caller choose what to do with it.  By
then, we would have readded the possibility as a feature,
_possibly_.  Then by having a From_ MBOXO quote filter as
a compile-time option.

But people complain.  So let us readd this almost as weird and
hacky as it always has been, and hope for the MIME and I/O layer
rewrite to bring this to a clean and more performant state, at
some future time, for which i look forward to.
Only allow this for the main body, and if that is text/plain.
And, as always, only if no privacy (S/MIME) signing is desired.