*v15-compat*: turn to value, extend meaning (imply `wysh')..
Change this from boolean to value so that we can extend the
meaning; henceforth, if set to a non-empty string, the command
modifier `wysh' will be implied for all commands where `wysh' can
be used to change from old to new shell-style argument quoting.

It should have been like that from the beginning. :(