Regulate shell expression parser <-> command argument parser..
[b28ad9f59] ("Fix" shell parser (/ command parser relationship)..,
2017-04-22) enabled us to handle

  define kb-clear {
    unbind * *;\
    bind base $'\n' mle-commit;\
    bind base $'\c?' mle-del-bwd;\
    bind base $'\cE' mle-go-end
  }

correctly: since we are not a shell with a syntax, but have to
parse a command in order to know how to treat the rest of the
line, and this results in some oddities.

For example in the above, the second argument of `unbind' needs to
be parsed over via dry-run (already beyond shell behaviour), and
that simply swallowed all the rest since .. how to handle
semicolon in a dry run?  Before the mentioned commit, it was not
recognized.

So that commit changed that with a "hack" in sofar as the shell
expression parser became extended and magically knew how to deal
with semicolon in a dry run (bad. and worse.).

And _this_ commit changes that to something regular.
All that is left to say has been said in the commit message of the
above commit already:

    All in all we have so many new possibilities that we really would
    need a complete review with those possibilities in mind, and much
    more testing.  A lot of things to do in the future.

Yep.  And much, much more to come.