XXX (BWDIC!) "~^head ins to|cc|bcc": verify *expandaddr* right away!..
Just like we do for ~t, ~c, ~b, immediately test whether the given
receiver complies to *expandaddr* settings.  Instead of waving
through names explicitly under the assumption that they are
aliases to be expanded/catched later on.

Like that the return value can be properly tested right away, also
for names, so that the return value indication is now equal
through all possible receiver types.  Add a test for that.

XXX We need a mechanism so that the `vput' prefix can be used
XXX for aliases.  But that does not cover *mta-aliases*!
XXX Hmm. Maybe pseudo headers to-name, cc-name, bcc-name?
XXX Offer a `vexpr' subcommand??