(BWDIC!) mx_termios_controller_setup(): reduce initial screen width (again)..
We now export $COLUMNS set to the real detected terminal width,
not to the possibly reduced value which we use ourselfs to not
cause wrapping when using the last column.  (On terminals where
this is true, thus always for us, without termcap support.)

On the other hand, do use the reduced size in batch mode.