Stop prefixing $features and $tls-features #, use , instead..
This actually is a fixup for [624fc223] (`if', `elif': support
wysh; add two argument -N, -n, -Z and -z, 2019-02-26), because
since then things like

      \if "$features" =% +key-bindings

will error out if *v15-compat* is set to a non-empty value

  ? \if "$features" =% +key-bindings
  mailx: `if' conditional: missing expression (premature end)

because in order to avoid that a leading + in $features becomes
*folder* expanded we placed the (then not) comment character
first.

Since these two strings are comma-separated lists, the simpled
solution seems to be to place an empty entry first.  Thus.