a_cwrite_save1(): FIX: ensure pre-v15 MBOX separation "in between" messages..
since messages will be written date sorted when doing eg "copy * TARGET",
and we do not care about the validity of the source, only the target,
it can be that a last entry in source MBOX is taken "as-is", without
a necessary separating line that is, and that thus the follow up
entry will be corrupted.

What a mess.
As a temporary workaround, ensure that the target mailbox is validated
after writing each and every message.

P.S.: we already have such a "add trailing \n" hack, in send.c:sendpart(),
but only for those cases where we are going to the display.
We cannot overload that occurrence, though, since it could put that \n
after each MIME part.
Model-View-Controller, and objects, this is where we MUST go to!!!

P.P.S.: test follows in a separate commit