[net-test] (BWDIC!) Extend authentication stuff, support automatic such..
In order to finally "just work" (tm), to reduce the number of
variables which need to be set to configure a single protocol, and
also to require lesser code once it is possible to do the
configuration also as part of an URL query string, rework the
authentication stuff a bit.

So with this we now have the framework to automatically handle
AUTHentication, by parsing those mechanisms which the server
supports, collate that with user wishes, and finally select the
best mechanism in the current context.

This also changes several other things:

- Add XOAUTH2 authentication (at least as a name).
  This is BWDIC!, because we falsely claimed OAUTHBEARER to be
  XOAUTH2 in the past.
- EXTERNAL and EXTERNANON require TLS.
- OAUTHBEARER and XOAUTH2 require TLS.