Fix *recipients-in-cc*: do not mix propagations..
If a receiver came in only via reply-to: or mail-followup-to: we
yet simply put her into the To: list "due to lack of a better
place".  This however looks odd if with *recipients-in-cc*
already addressed receivers are "declassified" into Cc:.

So we need a multi-pass approach instead, to look out whether To:
will be empty after the r-t / m-f-t has been taken over, or not.
Propagate any yet non-addressed to To: only if that is not empty,
otherwise stuff them into Cc:.  In conjunction with the other
address massages applied elsewhere this has the nice side effect
of finally placing in To: the one we really want there!