a_ignore_lookup(): FIX: perform exact testing..
woops, the complete rewrite of this in 2016 aided too much in code
savings, and has had the bug that cs_cmp_case_n and cs_cmp_n (in
effect) was used, but it was not checked whether the field storage
was NUL terminated at the length of the user buffer, too.
Therefore from_ and from matched equal, for example.