a_xtls_protocols[]: fail if proto is unavailable (Martynas Bendorius)..
If the SSL/TLS library does not offer mx_HAVE_XTLS_CONF_CTX
(SSL_CONF_cmd()) we have a builtin fallback for Min/MaxProtocol=
and Protocol= that iterates over this array to map a given user
string to a protocol.

Unsupported protocols will have their original SSL_OP_NO_xy and
xy_VERSION macros defined to 0 by us, so that the entry is always
available.  Until now, however, we then used that "0" directly as
arguments to SSL_CTX_set_m{ax,in}_proto_version() and maybe
SSL_CTX_set_options(), which is a bad idea except for
_max_proto_version().  Instead error out in these cases.
Thanks Martynas for pointing this out.