mx_child_fork(): unset knowledge of TTY ownership in child processes..
If a child would start another child through these functions
we yet would have acted as if we were the main program.
This is no good (but does not matter since we yet do not use
the functions recursively), so just loose the knowledge.

To make this work gracefully the fork()ed child would need
to go through a full environmental detection process, which
is too blown for our purpose here.  Maybe later on we move
the stuff to SU, and then a generic library init function,
which multiplexes initialization dependent on arguments,
will resetup the library state so that parent and child each
can work gracefully by themselves.