(BWDIC!) Add *mta-bcc-ok* (Kevin McCarthy)..
I fell out of the clouds ;) when he wrote on mutt-dev

 |Courier MTA doesn't remove the Bcc header when recipients are passed on
 |the command line.  I'm currently not in a position to verify this
 |behavior, so I'm assuming the ticket is correct.

 |It looks like Mutt provides $write_bcc, which allows the removal of the
 |Bcc header from the message.  However, I believe this will also remove
 |it from the Fcc copy.
 |
 |The ticket asks if there is a way to turn off passing the recipients on
 |the command line.  I'm wondering if this would be a generally useful
 |option.

Whereas we provide *mta-no-receiver-arguments* this does not help
people which use exim4 (which also behaves so) or Courier, let
alone by default.  Therefore change our default behaviour and
create a Bcc:-less version of messages in a_sendout_transfer()
before passing it through to MTAs, unless the new *mta-write-bcc*
is set.  This is backward incompatible.
We may be able to improve performance of this in v15.